+++ packages/hardhat/contracts/MockBridge.sol
@@ -9,21 +9,21 @@ import {IBridge} from "./interfaces/IBridge.sol";
 contract MockBridge is IBridge {
     event TokenTransferred(
         address indexed token,
-        uint32 indexed chainId,
+        uint64 indexed chainId,
         address indexed target,
         uint256 amount
     );
 
     event TokenTransferredWithData(
         address indexed token,
-        uint32 indexed chainId,
+        uint64 indexed chainId,
         address indexed target,
         uint256 amount,
         bytes data
     );
 
     event CrossChainCalled(
-        uint32 indexed chainId,
+        uint64 indexed chainId,
         address indexed target,
         bytes data
     );
@@ -35,7 +35,7 @@ contract MockBridge is IBridge {
     /// @param amount The amount of tokens to transfer
     function transferToken(
         IERC20Metadata token,
-        uint32 chainId,
+        uint64 chainId,
         address target,
         uint256 amount
     ) external override {
@@ -50,7 +50,7 @@ contract MockBridge is IBridge {
     /// @param data Additional data to send with the transfer
     function transferTokenAndData(
         IERC20Metadata token,
-        uint32 chainId,
+        uint64 chainId,
         address target,
         uint256 amount,
         bytes calldata data
@@ -63,7 +63,7 @@ contract MockBridge is IBridge {
     /// @param target The target contract address on the destination chain
     /// @param data The calldata to send to the target contract
     function callCrossChain(
-        uint32 chainId,
+        uint64 chainId,
         address target,
         bytes calldata data
     ) external override {
