+++ packages/hardhat/contracts/interfaces/IRiskHub.sol
@@ -4,6 +4,6 @@ pragma solidity ^0.8.10;
 import {IERC20Metadata} from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
 
 interface IRiskHub {
-  function tradeFromDex(uint32 chainId, uint40 timestamp, IERC20Metadata tokenIn, uint256 amountIn, uint256 amountOut)
+  function tradeFromDex(uint64 chainId, uint40 timestamp, IERC20Metadata tokenIn, uint256 amountIn, uint256 amountOut)
   external;
 }
